stuff
mcd: U_BASE, bigWall block deathtile change from floor[0] to object[3]

SellState: fix colors from Debriefing, etc.
DebriefingState: lots of debug
	- don't count unresearched corpses as artefacts
Map: fix CTD
TileEngine::detonate() - no smoke if objectpart blocks smoke
TextEdit: functions for reverting savename before committed to disk, etc.
ListGamesState, ListSaveState: allow reverting savename before commit to disk
	- note: uses Escape and keyCancel, but the latter
		could be set to a different key
BattleUnit: deathSound now defaults to -1

+++ src/Interface/TextList.cpp
@@ -1030,7 +1030,7 @@ void TextList::clearList()
 /**
  * Scrolls the text in the list up by one row or to the top.
  * @param toMax			- true scrolls to the top of the list; false => one row up
- * @param scrollByWheel	- true uses wheel scroll
+ * @param scrollByWheel	- true uses wheel scroll (default false)
  */
 void TextList::scrollUp(
 		bool toMax,
@@ -1059,7 +1059,7 @@ void TextList::scrollUp(
 /**
  * Scrolls the text in the list down by one row or to the bottom.
  * @param toMax			- true to scroll to the bottom of the list; false = one row down
- * @param scrollByWheel	- true uses wheel scroll
+ * @param scrollByWheel	- true uses wheel scroll (default false)
  */
 void TextList::scrollDown(
 		bool toMax,
@@ -1076,7 +1076,7 @@ void TextList::scrollDown(
 		else
 		{
 			if (scrollByWheel == true)
-				scrollTo(_scroll + Options::mousewheelSpeed);
+				scrollTo(_scroll + static_cast<size_t>(Options::mousewheelSpeed));
 			else
 				scrollTo(_scroll + 1);
 		}
@@ -1120,7 +1120,7 @@ void TextList::updateVisible()
 
 /**
  * Changes whether the list can be scrolled.
- * @param scrolling - true allows scrolling
+ * @param scrolling - true allows scrolling (default true)
  * @param scrollPos - custom +/- x_offset for the scroll buttons (default 0)
  */
 void TextList::setScrollable(
